requirements.txt
dash>=2.14.1
numpy>=1.26.1
pandas>=2.1.2
openpyxl>=3.0.7
dash-bootstrap-components>=1.5.0
dash-core-components>=2.0.0
dash-html-components>=2.0.0
dash-table>=5.0.0
dash-daq>=0.5.0
pdfkit>=1.0.0
dash-iconify>=0.1.0
python-dotenv>=1.0.0